# Conflicts:
#	src/main/java/at/qe/skeleton/models/SensorStation.java
#	src/main/java/at/qe/skeleton/models/Userx.java
#	src/main/java/at/qe/skeleton/models/demo/LogEntry.java
#	src/main/java/at/qe/skeleton/models/demo/LogEntryType.java
#	src/main/java/at/qe/skeleton/models/demo/Message.java
#	src/main/java/at/qe/skeleton/models/demo/UserStatus.java
#	src/main/java/at/qe/skeleton/models/demo/UserStatusInfo.java
#	src/main/java/at/qe/skeleton/services/UserService.java
#	src/test/java/at/qe/skeleton/tests/EqualsImplementationTest.java

+++ src/main/java/at/qe/skeleton/models/AuditLogEntry.java
@@ -11,7 +11,7 @@ public class AuditLogEntry {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
     @Column(name = "TIMESTAMP", nullable = false)
     private LocalDateTime timestamp;
     @Column(name = "LOG_LEVEL")
@@ -24,7 +24,7 @@ public class AuditLogEntry {
     @Column(name = "DETAILS")
     private String details;
 
-    public AuditLogEntry(Long id, LocalDateTime timestamp, LogLevel level, LogEntity origin, String details) {
+    public AuditLogEntry(Integer id, LocalDateTime timestamp, LogLevel level, LogEntity origin, String details) {
         this.id = id;
         this.timestamp = timestamp;
         this.level = level;
@@ -36,7 +36,7 @@ public class AuditLogEntry {
 
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
