+++ src/main/java/at/qe/skeleton/repositories/AccessPointRepository.java
@@ -4,5 +4,5 @@ import at.qe.skeleton.models.AccessPoint;
 
 public interface AccessPointRepository extends AbstractRepository<AccessPoint, Long> {
 
-    AccessPoint findFirstById(Long id);
+    AccessPoint findFirstById(Integer id);
 }
