+++ src/main/java/at/qe/skeleton/repositories/SensorStationRepository.java
@@ -4,5 +4,5 @@ import at.qe.skeleton.models.SensorStation;
 
 public interface SensorStationRepository extends AbstractRepository<SensorStation, Long>{
 
-    SensorStation findFirstById(Long id);
+    SensorStation findFirstById(Integer id);
 }
