+++ src/main/java/at/qe/skeleton/repositories/SensorValuesRepository.java
@@ -4,5 +4,5 @@ import at.qe.skeleton.models.SensorValues;
 
 public interface SensorValuesRepository extends AbstractRepository<SensorValues, Long> {
 
-    SensorValues findFirstById(Long id);
+    SensorValues findFirstById(Integer id);
 }
